GH-1366 Refactored dashboard page structure and extracted shared dashboard components
