+++ master/front-end/src/components/dashboard/DashboardCard/styles.module.css
