+++ master/front-end/src/components/dashboard/DashboardGauge/ArcSegment/index.tsx
