+++ master/front-end/src/components/dashboard/DashboardGauge/StatCards/index.tsx
