+++ master/front-end/src/components/dashboard/DashboardGauge/StatCards/styles.module.css
