+++ master/front-end/src/components/dashboard/DashboardGauge/index.tsx
