+++ master/front-end/src/components/dashboard/DashboardGauge/styles.module.css
