+++ master/front-end/src/components/dashboard/DashboardPagesFirstSection/index.tsx
