+++ master/front-end/src/components/dashboard/DashboardPagesFirstSection/styles.module.css
