+++ master/front-end/src/components/index.ts
@@ -15,14 +15,15 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
+export { DashboardPagesFirstSection } from "./dashboard/DashboardPagesFirstSection";
 export { NoRequiredAuthorityTooltip } from "./NoRequiredAuthorityTooltip";
-export { DashboardPagesFirstSection } from "./DashboardPagesFirstSection";
+export { DashboardGauge } from "./dashboard/DashboardGauge";
+export { DashboardCard } from "./dashboard/DashboardCard";
 export { LanguageSwitcher } from "./LanguageSwitcher";
 export { TooltipWithCopy } from "./TooltipWithCopy";
 export { LinearProgress } from "./LinearProgress";
 export { UniversalModal } from "./UniversalModal";
 export { AccessProvider } from "./AccessProvider";
-export { DashboardCard } from "./DashboardCard";
 export { HashNavigable } from "./HashNavigable";
 export { UserRoleTags } from "./UsersRoleTags";
 export { EmptyHandler } from "./EmptyHandler";
