+++ master/front-end/src/pages/Dashboard/DashboardJava/DashboardLegendItem/styles.module.css
