+++ master/front-end/src/pages/Dashboard/DashboardPersistence/DashboardPersistenceLegend/index.tsx
